+++ drivers/net/usb/smsc75xx.c
@@ -1460,7 +1460,7 @@ static int smsc75xx_bind(struct usbnet *dev, struct usb_interface *intf)
 		return ret;
 	}
 
-	dev->data[0] = (unsigned long)kzalloc(sizeof(struct smsc75xx_priv),
+	dev->data[0] = (uintptr_t)kzalloc(sizeof(struct smsc75xx_priv),
 					      GFP_KERNEL);
 
 	pdata = (struct smsc75xx_priv *)(dev->data[0]);
